Added new configurtaion option for hiding dot files.
When new option "hide_dot_files" is set to 1, every file or directory beginning
with a "." is not shown in the directory listing. But you can still access the
file / directory by typing its name into the "Go to directory/file" box.

This option is also available in the user-dependent configuration.