Changed usage of template class.
We don't use the "new" (2006...) set_var() (and so on) methods any more,
because it causes some problems concercing variable substitution. Instead we
use fillin() and so on and start loop parsing manually using parse_loop(). I
think there is actually no need of using the new methods. Maybe we can switch
it later.